requirements.txt
openai-whisper
torch
transformers
sentencepiece
sacremoses
tts
soundfile
gtts
streamlit
SpeechRecognition
PyAudio
